requirements.txt
requests
pymongo==4.5.0
pillow==10.2.0